Fixed errors when loading bad data from the DB (#3)
The update to use Image constructor instead of URL for icon has broken out
storage of the data, resulting in bad data in the DB and errors when
rehydrating it. Long term I would like to move back to using URL for the `icon`
property as this is simpler to work with.