Fixed tests not running as expected for `yarn test` command
When executing tests via `yarn test` they are executed inside of a
container. The tests that are being executed are the ones added when the
container was first built. Subsequent changes to the tests are not causing
docker-compose to rebuild the container meaning the same set of tests were
always being executed. This change mounts the `src` directory into the
container so that the tests are always up to date.