Updated recipients for outgoing activities
The `to` field for Create<Article> activities was incorrectly set to the
followers collection, but the content is intended to be public. This may have
been causing issues with users being able to see the content depending on if
they are following the actor.

What we've fixed here is made public activities addressed `to` the public
collection, and adding a `cc` to the followers because they are also an
intended recipient, though not the primary.