Used exec to run tests in existing container
The idea here is to simplify and streamline the dev stack. So rather than a
separate docker-compose file for testing, and spinning up new containers we can
run the tests in the existing activitypub container we use, which also has the
local src files mounted.

`yarn dev` will start the stack as usual, but now it does it as a daemon, which
means it runs in the background. We also pass the --no-recreate flag which
ensures that running it multiple times doesn't destroy the container and is
essentially idempotent.

`yarn stop` will bring the stack down

`yarn logs` will tail the activitypub logs for the container

`yarn test` uses the `exec` docker compose command to run the tests, and also
calls `yarn dev` first - which makes sure the stack is running if it wasn't
already.

`yarn fix` will remove local containers for activitypub and nginx and then rebuild them