Added Zod for type coercion and validation
There is a bug where the incorrect type was being passed as the `published`
property when creating an instance of Article. This should have been caught by
the type system, but we were dealing with implicit `any` types for the body of
the request. This change introduces the use of Zod for parsing the request body
as well as providing types so that we can catch issues like this in future.