Allowed connections to private addressed when testing
This allows fedify to work with our docker setup for testing. We use a wiremock
container to represent external activitypub services, but when fedify attempts
to make requests to this it will error as the hostname resolves to a private ip
address, setting this flag allows us to bypass this security restriction.