Added initial reply action
ref https://linear.app/tryghost/issue/AP-306

We force a mention for the author of the object being replied to in
order to make sure they get a notification in mastodon, and we also
make sure that the activity is sent directly to the author, in case
they are not one of our followers.

The endpoint responds with the created activity so that the client can
update its local cache with the new object.