Added dedicated endpoint for retrieving activities for a given actor (#40)
* Added dedicated endpoint for retrieving activities for a given actor

refs [AP-377](https://linear.app/tryghost/issue/AP-377/inbox-returning-33mb-of-data)

Added a new endpoint for retrieving activities for a given actor:

```
GET /.ghost/activitypub/activities/:handle?limit=<number>&cursor=<string>
```

At the moment this returns the same data as the inbox handler but in the future
we are likely to want to augment this with additional data (i.e including the
actor's own activities). This endpoint is designed to be used by the activity pub
UI in Ghost Admin and not other activity pub compliant servers / clients.

The returned data is paginated using cursor based pagination. The `cursor` query
parameter is optional and is used to fetch the next page of activities. The `limit`
query parameter is also optional and is used to limit the number of activities
returned in a single request.