Wired up authentication to tests
refs https://linear.app/tryghost/issue/AP-288

This makes sure that all of our tests use authentication. We have to stop
communicating with the `activitypub-testing` service directly, because we need
to replicate how this runs in production. How this generally runs is that there
will be a host `example.com` and everything on `/.ghost/activitypub` will hit
our service, and everything else will hit a Ghost instance - specifically for
us we need the `/ghost/.well-known/jwks.json` endpoint to be reachable for auth
to work.

We have two wiremock instances, `fake-external-activitypub` is used
to mock external services, so we can test things like follow requests and
sending to external inboxes. `fake-ghost-activitypub` is the actual service
under test, it proxies everything to our underlying `activitypub-testing`
container, but allows us to add mocks for the jwks.json file. This essentially
behaves as the `nginx` container in dev, or fastly in production.

Listening on port 80 gives nicer urls

Hardcoded keypair saves generating the keys, and allows us to cache them in the db.