Ensured consistent URL encoding in the endpoints
no refs

Updated the API endpoints to consistently use the same method of URL encoding (
previously we used `encodeURIComponent` in some places and base64 encoding in
others). This change ensures that all endpoints use `encodeURIComponent` to encode
URL parameters.

This change also normalises the naming of the `cursor` parameter in the
endpoints - `cursor` and `next` are now used consistently across all endpoints