Disabled automatic creation of entries in the sites table
ref https://linear.app/ghost/issue/AP-500

Rather than having all requests create an entry in the sites table if it's
missing, we want to move to a model where an entry in the sites table
determines whether a site can use the service or not. This will tie in with the
onboarding work in Ghost where a Ghost site will essentially create the entry
in the sites table, and then configure the activitypub integration locally

We also have to move the ping middleware, so that it's not caught my the missing
entry in sites table for things like docker healthchecks.