Retried fetching JWKS when authenticating requests
ref https://linear.app/ghost/issue/AP-500

Some authenticated requests come into ActivityPub right on the cusp of when
Ghost is booted and ready - which means the request for the JWKS can fail!

You might think that this would be alleviated by caching our JWKS - which is
something we should definitely do, but in the case where the site has booted
yet it is generally the first request anyway, so a cache wouldn't fix that.