Added support for configuring log level with env
ref https://linear.app/ghost/issue/AP-551

We are unable to replicate some issues locally and would like to enable more
specific logs in production, rather than having to change the hardcoded values
we should be able to do this via the enviroment.

This adds support for three new environment variables:

  - `LOG_LEVEL_FEDIFY` sets the log level for internal fedify logs
  - `LOG_LEVEL_ACTIVITYPUB` sets the log level for our logs
  - `LOG_LEVEL` is a fallback if either of the above are not set

If none of these are set we continue to default to the existing hardcoded values.