Updated query for retrieving thread parents / children (#163)
refs [AP-579](https://linear.app/ghost/issue/AP-579/investigate-parent-post-visibility-issues-in-replies-to-articles)

Updated the query for retrieving thread parents / children so that it takes into
account the activity type. This is needed because when we query on the object
it is possible for there to be multiple activities with a matching object if
the actvity embeds the object (i.e in the case of a `Like` activity). This is
important when looking up the immediate parent of an activity (as we only expect
there to be 1 parent). Scoping the type of activity should be fine for our
use case of "threads" (a list of comments)