Ensured scoped kvstore is reinitialized on each incoming message (#190)
no refs

When using the message queue in production we was finding that seemingly
randomly, activities would be attributed to the the wrong site. We narrowed
this down to the fact that the scoped kvstore was not being reinitialized
on each incoming message. Fedify reuses contexts when queueing messages so its
possible that the already initialized scoped kvstore was being used for a
different site. This change ensures that the scoped kvstore gets reinitialized
on each incoming message.