Fixed loading profiles for users from WordPress plugin
ref https://linear.app/ghost/issue/AP-597

There was a bug in fedify when loading the outbox for an Actor if the outbox
was the type of OrderedCollectionPage. This was causing our code to throw an
error when loading their profile as we would fetch their outbox to get the
recent activities. This bug was fixed in v1.2.6 of fedify.