Added support for updating the site actor
ref https://linear.app/ghost/issue/AP-592

When we fetch the site data we'll attempt to update the site actor too, this
will ensure that when Ghost boots any changes are propagated to us.

We're handling the sending of the Update activity ourself, and there's room to
remove duplication here. I'd like to think about what pattern we use for this.
Ideally I think a repository would be responsible for this or at least abstract
it.