Added support for updating the site actor
closes https://linear.app/ghost/issue/AP-592
closes https://linear.app/ghost/issue/AP-637

We've refactored the update of the site actor so that it can be used in multiple places.
On top of that we've refactored fetching user data from the database, and setting it,
and all places which need this data use the helpers so we can ensure that all data stored
is complete. This fixes the bug AP-637 which was caused by incomplete user data stored
in the database.

With the refactored update of the site actor we're able to include updates when we fetch site
settings, which happens on every boot of Ghost - this ensures that newly onboarded sites
as existing sites will always update their actor to have the latest data on boot.