Handle authorized fetch for announced objects
Some ActivityPub implementions, threads.net as one example, require
the use of authorized fetch to read objects, even if they are public.

When we use the `getObject` method of the Activity class the request
is made without authorized fetch, which causes the handler to error.

Our lookupObject helper handles the authorized fetch component, which
fixes the error.