Updated application to use `isFollowedByDefaultSiteAccount`
no refs

Updated application to use `isFollowedByDefaultSiteAccount` to determine if an
actor is followed by the default site account. This should replace the old
`isFollowing` helper function that reads from the old KV store.

Affected areas:

- Search results to indicate if a search result account is followed by the
  default site account
- Profile result to indicate if a profile is followed by the default
  site account
- Follows result to indicate if a profile follow result is followed by the
  default site account
- `Create` handler when checking if an item should be added to the inbox
- `Announce` handler when checking if an item should be added to the inbox
- `AnnouncedCreate` handler when checking if an item should processed

This change also ensures that `site` on the `AppContext` is typed correctly so
that request handlers do not have to manually resolve the site via the site service
(as `site` gets set in a middleware already)