Removed duplicate method for fetching default account
Whilst we were building out the infrastructure for accounts and sites we ended
up with some duplicate code, this consolidates it into the account service,
slightly updating the signature to never resolve with null, throwing if an
account is not found instead. We also clean up all the call sites and tests.