Added DB tables for storing and interating with posts
refs https://linear.app/ghost/issue/AP-714

We want to move away from storing raw ActivityPub objects in a KV store and
instead start storing the data in the format which best suits our usecases.

This means moving to a new schema where we're able to store our concept of a
post and use joins to be able to look up information such as who has liked or
reposted a post, as well as which posts belong in a users feed.

This schema is the result of some long term performance testing as well as
audits into our UI and the data it needs in order to render.