Added support for liking to the Post entity
refs https://linear.app/ghost/issue/AP-728

This adds the initial support for liking a Post. It's a bit tricky because of
how messy AP is - we need to be able handle duplicate likes for example we may
be forwarded multiple of the same Like or have likes come in via the Inbox as
well as our internal API's.

We also need to work out how to handle the like count - whether we increment
that in place on the model, which could be wrong. Or just update it on
persistence, which is more accurate but then do we need to mutate the entity in
the save method? We can work this out as we go, with a federated network we're
not expected to have exact numbers for these things.