Wired up the LikeHandler to the db
ref https://linear.app/ghost/issue/AP-728

We've had to update the Like activity dispatcher to take dependencies in the
same "create" style as the others. The existing logic hasn't changed, but we've
added a few lines which will interact with the Post and Account entities,
registering a like, and persisting it to the database.

We've also wired up FedifyContextFactory at the dispatcher level, so this is
usable for all dispatchers going forwards.