Fixed tests dealing with Like activities
ref https://linear.app/ghost/issue/AP-728

Our tests dealing with like activities were calling the like action with
activities rather than objects, and this caused an error in the new code.

The error is correct, we do not support liking a Create activity, but it wasn't
caught in the existing code. We update the feature files and step defs to make
it easier to refer to the specific activity or object.