Handled constraint conflicts when writing to feeds
ref https://linear.app/ghost/issue/AP-765

With the new UNIQUE constraint on the feeds table, we need to make sure we
handle conflicts when inserting. The `batchInsert` method in knex is just a for
loop like this, but doesn't allow using the `onConflict` method. We've pulled
that out and ensured everything is still run inside of a transaction