Improved error handling for getByApId methods
closes https://linear.app/ghost/issue/AP-774
closes https://linear.app/ghost/issue/AP-771

We were incorrectly throwing a URL instead of an Error when the Actor wasn't
found. This has been fixed, as well as the error message improved so that we
can see exactly what the problem was.

We were throwing when recieving invalid posts, but this is not an error, it's
an expected path - so instead we're returning null. We've also improved the
error message for when the post isn't found at all.