Updated Post to support creating replies
Rather than store a Post object and need to make a lot of queries to populate
Post's with their reply data we just store the id instead.

This means we can actually populate those fields from the repository now,
rather than hardcoding them to null.

We also have the threadRoot logic in the entity as this is business logic.

We also allow passing the `apId` because that is used by the publishing service
when publishing Posts.