Wired up like action to the new db tables
closes https://linear.app/ghost/issue/AP-724

For outgoing likes, we want to store the like in our `likes` table immediately,
this is so that we can track the likes of posts outside of our system, as well
as not relying on the incoming activity to store likes for posts inside out
system. This will allow us to eventually optimise and not send outgoing
activities to ourself, instead updating state immediately, as we do here.