Added suport for removing like from post
ref https://linear.app/ghost/issue/AP-731

Before we wire anything up to the unlike action, we need to addthe capability
of unliking to our model. Now that we have likes and unlikes we need to make
sure that they stay in sync, so we remove from the opposite list when
performing an action, to cancel out any opposing actions.

With this approach we can no longer have a one-sided get method, because it
would then be possible for things to fall out of sync, so we move this to a
getChangedLikes method, which will read from and clear both lists. It is still
up to the repository to make sure it actually deals with these correctly ofc.