Updated the repository to handle removing likes
ref https://linear.app/ghost/issue/AP-731

Here we actually do the work to remove the rows from the likes table. We can
safely ignore this in the case of a new post because it's impossible for there
to already be rows. Knex will return the deleted number from a call to `del` so
we're able to so a single update for the count, taking into account the
inserted and deleted rows.