Wired up unlike action to the new db tables
closes https://linear.app/ghost/issue/AP-731

For outgoing unlikes, we want to remove the row from our `likes` table
immediately, this is so that we can track the removal of likes of posts outside
of our system, as well as not relying on the incoming activity to remove likes
for posts inside our system.

This will allow us to eventually optimise and not send outgoing activities to
ourself, instead updating state immediately, as we do here.