Removed duplicate articles from feed
ref https://linear.app/ghost/issue/AP-791

Now that we've fixed newly created articles we want to clean up the old ones.
The rows that have a missing `excerpt` are the ones which have the correct
`uuid` so we want to retain these rows, and copy over the missing excerpt.

First we mark all of the rows _without_ a missing excerpt with a `type` of 3,
this is currently unused in the system and we use it as a mark to delete them.

Then we copy the excerpt from posts with an excerpt to ones without, we do this
only when the `title` is the same and the `type` is 1 (an article).

After we've copied the excerpt, we delete the previously marked posts.