Added a Post -> PostDTO serialiser
ref https://linear.app/ghost/issue/AP-723
ref https://linear.app/ghost/issue/AP-722

This will be required by both the note and the reply HTTP API's so that they
can return the correct response type. Note that we currently use the `apId` as
the `id` to follow how the FeedService is implemented, this is to decouple the
changes, as other endpoints rely on the id being a URL. Once everything is
migrated to use the new DB we can then switch to using uuids here.