Added content cleaning to Post
closes https://linear.app/ghost/issue/AP-792

We were storing posts without cleaning their html first, which meant we were
storing the wrong content, as that was not what is in the activity.

The cleaning of content is business logic and so belong in the entity. Whilst
the publishing is not hooked up to the entity we've had to duplicate this logic
for now, but eventually ti will be removed from the publishing service, which
will hook into our events to create an Article when the post is persisted.