Added deleted column to posts table
ref https://linear.app/ghost/issue/AP-812

We want to use soft deletions for posts, this gives us flexibility for
reverting delete operations on our content in the future, it also
allows us to be more efficient when reading posts.

When attempting to retrieve a post by the ActivityPub ID we read from
our own database, and if it's not there we attempt to fetch it over
the network.

For a deleted post, if we were to remove it completely from our
system, we would always have to do a network request to find a 404.

Instead, if we keep the post locally and replace it with a Tombstone,
we don't have to hit the network, we know that it is deleted and can
act accordingly.

Seeing as we're not currently filtering by the `deleted` column we
haven't added an index for it yet.