Switched from deleted column to deleted_at
ref https://linear.app/ghost/issue/AP-812

After some discussion in Slack we want to use a deleted_at column
rather than a deleted one. This gives us the added benefit of knowing
when a delete happened, which can be useful when debugging. The trade
off here is that the column is larger, but given that we don't need to
index the `deleted_at` column at the moment, we can live with it.