Added delete method to Post entity
ref https://linear.app/ghost/issue/AP-812

We've kept the `deleted` property private for now, and expose access
via a static `isDeleted` method on the class. The idea here is that
the soft deletion is an implementation detail which we're not sure
we want to leak to consumers right now.

When a Post is deleted, we make sure that none of the content is
available from the entity - this means we can be sure that we never
serve deleted content to the client. Generally we shouldn't be dealing
with instances of deleted entities - but there are a few cases we might:

1. When we've just called `delete` on it
2. When we're working with reply threads and a post in the chain is
   deleted, we may want to indicate that there _was_ a parent a
   Tombstone post with no content gives us the flexibility to do that.

The addition of the Tombstone type to PostType meant that we had a few
issues when working with the PostType because we don't want to allow
creating posts as a Tombstone, and when we call for example the
`getApIdForPost` method we don't want to be dealing with Tombstones
either so we've introduced a new CreatePostType which is everything in
PostType *except* the Tombstone