Support persistence of deleted posts
ref https://linear.app/ghost/issue/AP-812

We need to mark deleted posts as deleted in the DB, as well as emit a
PostDeletedEvent so that other systems can react to it, for example
removing the post from the feed after a deletion.

Newly created posts that are immediately deleted are not a use-case we
have and it's not clear on how that should work, this implementation
simply does not persist them for now, but we can change that in future
if we need to support this.

When a post is deleted, the only column we write to is the deleted_at
column, this ensures that the content is still available for audit,
debug, undoing etc...

We also update the reply count of its parent, if it exists.

We also have to add a new check that the Post author Account has an id
property, I would like to move this constraint to the entity layer,
but that is outside the scope of this PR. Right now this check still
happens, but it is implicit at the DB level with a foreign key
constraint on the column - this just makes it explicit so that 1) we
are all aware of it and 2) TypeScript doesn't complain about
referencing author.id when instantiating the PostDeletedEvent.