Made all test:* scripts runnable from the host
Some of our test scripts required to be run inside of docker which
makes working with them locally really annoying.

This standardises all of our test scripts using the `_test` prefix for
the command which executes the test and the `test` prefix for the
command which can run the test from the host.

This means we can use `yarn test:unit` or `yarn test:integration` and
they will run in the docker environment with access to everything they need.

We are now running the unit tests inside of docker just to be
consistent, I haven't seen a noticable impact on the speed here - but
it can always be optimised in future if desired.