Supported running the test suite in parallel
Our integration tests were all using the same database, which meant
that they would interfere with one another and cause random failures,
we initially fixed this by forcing the tests to run in series, but as
our suite has grown this has caused it to slow down considerably.

This introduces a new test helper `createTestDb` which will create a
brand new database with the same schema as the activitypub one, so
that test suites can each create their own isolated environment.

We've had to use the `root` user for MySQL in our testing container so
that it has permission to create new databases.