Moved content preparation to the Post entity
ref https://linear.app/ghost/issue/AP-869

Our replies were not going through the publishing service which meant that
content preparation was not being applied to them. We have instead moved this
logic into the entity. And created two new static methods for creating Notes
and Replies for our internal accounts. This allows us to contain all of the
business logic in the enitity and ensure that it is applied consistently

We have also updated the reply action handler to use the `content` and `apId`
from the Post entity, in the same way that the note action does. This delegates
the business logic to our entity layer, keeping everything in one place.