Ensured the sanitisation of content
ref https://linear.app/ghost/issue/AP-870

We want to be sure that we're sanitising content when we read/write to the db
through the entity layer. But in the same pattern as the activities endpoint we
are also adding sanitisation to the feed service to ensure that any issue with
the storing of html content is still protected when reading data for the feed.