Updated repository to wait for events to process
closes https://linear.app/ghost/issue/AP-859

We don't want to run the event handlers outside the scope of a request, because
the container can be shut down once the request has finished executing. This
change keeps our event handlers running within the lifetime of the request.

Long term this will be handled by using a queue for the domain event system, so
they can be pushed outside of the request lifecycle and safely handled async