Removed likes for deleted posts
ref https://linear.app/ghost/issue/AP-864

When posts are deleted their likes are kept in the `likes` table,
which means that we fetch deleted posts when looking up a users liked
posts as well as we counts including likes for deleted posts.

In order to get liked posts for a user which don't include the deleted
ones we need to filter on the joined posts table - which is inefficient.

This also includes a migration to remove existing likes for deleted
posts. This migration is not reversible.