Converted URLs in Notes to links
closes https://linear.app/ghost/issue/AP-923

When we create notes and replies we want any URLs to be clickable links for
clients, both in Ghost and in 3rd party clients like mastodon. In order to
ensure this we convert the URL to an anchor tag, which is then used both
internally in Ghost & sent as the content over the ActivityPub network.

The first implementation of this missed the `createReply` method - so to
prevent problems like that in the future, I've made all of the options to the
ContentPreparer.prepare method required - this will force all call sites to be
updated when the options are changed. Making sure that we update all the
necessary places.