Handled missing member content marker in ContentPreparer
ref https://linear.app/ghost/issue/AP-961

The nuance of setting the content to an empty string if the prepared content
was the same as the input when it was a non-public post was all handled in the
entity, and it becomes confusing when we add extra content preparation steps.

This small refactor updates it so that we can immediately exit if the post
we're dealing with is private. It does mean that you should only ever pass the
removeMemberContent flag when dealing with a private post - but that is only
ever done in this factory function, and is covered by unit tests.