Handled completely private posts in ContentPreparer
ref https://linear.app/ghost/issue/AP-961

When we do `if (content === post.html) content = ''` we're secretly doing the
removal of member content, for the case where the whole post is member content.

When adding extra steps to the content preparation we may do other
modifications to the content which should not be done if the post
is members only.

Instead the method to remove members content will now remove everything if a
members content marker isn't found. This means that it should never be called
on a public post, this is covered by unit tests in our Post entity.