Updated getPostsByAccount query to get if a post id liked/reposted by the current user (#488)
Ref https://linear.app/ghost/issue/AP-1024/

Previously, this query (`getPostsByAccount`) was used to fetch only the posts of the current user for the Profile view, so `likedByMe` and `repostedByMe` indicated whether the posts were liked or reposted by that account. Now, this query will also fetch posts from other ghost accounts, so the `likedByMe` and `repostedByMe` fields should indicate whether the current user (the one viewing the profile) has liked or reposted the posts.