Added various counts onto the `Account` entity
no ref

Added the following counts onto the `Account` entity:

- `postCount`
- `repostCount`
- `likedPostCount`
- `followerCount`
- `followingCount`

Having these counts on the entity means we can move away from using "helper"
methods in services / repositories to retrieve this information from the
database. This is particulary useful when we retrieve a "remote" account and we
can't utilise the database to retrieve these count values (as we don't have
them). This will later come into play when we refactor the account service to
fetch account data either locally or remotely and return a single `Account`
entity that can be serialised by the controller action.