Handled inserting duplicate accounts
ref https://linear.app/ghost/issue/AP-987

We attempt to insert duplicate accounts due to a race condition between looking
for if an account exists and then trying to create it. Whilst we've stopped the
duplicate rows appearing in the database, we need to actually handle the error.

We can ignore the error and return the AccountType as if the insert succeeded
with the current usage. When we move the writing to the repository we will
return a duplicate error, and allow the caller to choose to fetch or not.