Fixed key_value table definition in migrations
ref https://linear.app/ghost/issue/AP-1055

When trying to writing a breaking test for errors with concurrent writes to the
key_value table I discovered that our migration was out of sync with our
production db definition, which was created before we had migrations in place.

The key difference (hehe) is that the `key` column length was too long, and
also missing a unique constraint.

This commit now correctly shows the concurrent write test as failing.