Fixed error for concurrent writes to key_value
ref https://linear.app/ghost/issue/AP-1055

A transaction doesn't help us when the row doesn't exist! Instead we can use
MySQLs ON CONFLICT and then merge the values, effectively doing an upsert.

Knex.js does provide an `upsert` method but this is only implemented for MySQL
and CockroachDB - wheras the onConflict works for SQLite and PostgreSQL, giving
us more portability of the code for an inherently uncertain future.