Added initial Result implementation
We want to stop throwing Errors and instead returning Results from our
methods in an effor to force us to handle errors explicitly and
correctly. This is because we're working in a decentralised system
where errors are abundant and expected, and we want to make them first
class citizens of our application.

The current state of the system of throwing errors is leading to a lot
of 500 errors in production, most of which can and should be handled
at the application level, rather than bubbling up.

This implementation is very small and essentially boils down to 2
element tuples containing either null and the value, or an error and null.

A few helper functions have been provided to make the DX a little
better, with the idea being that we can either build this up, or
replace it with a library once we have a better understanding of our
use-cases and desired feature set.