Added `exhaustiveCheck` helper
This is to reduce the boilerplate when writing switch statements to handle
Results which are Errors. The `default` case of all switch statements should
return a call to the exhaustiveCheck function, passing the error to it. This
will provide use with type safety that we have handled all possible cases.