Created docker network & configured aliases for tests
ref https://linear.app/ghost/issue/AP-963
ref b68297110e8108676a975fd90bcd9078a72191b2

Our WebFinger implementation validates that the incoming `acct` resource has a
proper hostname, and we require there be a top level domain. In testing our
hostnames were all mapped to docker compose service names and so they do not
have a `.` in them and don't pass this test. This was also the case in the
linked commits `isHandle` function which has a special branch for test mode.

With docker compose we are able to specify aliases for services on a particular
network. I've gone for a separate testing network here to completely isolate
our testing stack, which is probably unecessary, especially because we're using
the `.test` TLD which shouldn't interfere with our dev stack, but I think it's
cleaner to have these on separate networks anyway and avoid any cross pollution.