Stripped dubdubdub prefix from handles
ref https://linear.app/ghost/issue/AP-963

All of the handles generated by us now no longer have a www. in the host. By
updating this util we ensure that handles throughout the application have this
www. prefix stripped.

This includes handles external to the our service, which means we may show
incorrect handles in the UI for external sites, however we've deemed this to be
an edge case for now, and in the case of a bug, it's purely a cosmetic one,
those account can still be searched and interacted with using their handle.