Used `index` for AP URL generation
ref https://linear.app/ghost/issue/AP-902

Since we now allow usernames to not be 'index' - but we want to
maintain 'index' as the identitifier in the path - we need to hardcode
these URL's to use 'index' in them. When we eventually have multiple
users per site, we will switch to the UUID for the identitifier in the path.

For now we want a single codepath, so we don't need to switch on index
vs uuid or index vs someother username. This keeps things simple and
easier to reason about when we do finally migrate to multiple users.